+++ js/validForm.js
@@ -1,6 +1,18 @@
+import {isEscButton} from './utils.js'
 const form = document.getElementById('upload-select-image');
 const hashtag = /^#[a-zа-яё0-9]{1,19}$/i;
 const pristine = new Pristine(form);
+
+
+//ESC
+const onPopupEscKeydown = (evt) => {
+  if (isEscButton(evt)) {
+    evt.preventDefault();
+    hideSucsessWindow();
+  }
+};
+
+
 
 function isHashtagValid() {
   const input = document.querySelector('.text__hashtags').value;
@@ -30,24 +42,39 @@ function isCommentValid() {
 pristine.addValidator(form.querySelector('.text__description'),isCommentValid);
 
 
+const hideSucsessWindow = () => {
+  const success = document.querySelector('.success');
+  success.remove();
+  window.removeEventListener('keydown', onPopupEscKeydown);
+};
+
+
 form.addEventListener('submit', (evt) => {
   evt.preventDefault();
 
   const valid = pristine.validate();
   if (valid) {
-    const formData = new FormData(evt.target);
-    fetch('https://27.javascript.pages.academy/kekstagram',
-      {
-        method: 'POST',
-        body: formData,
-      });
-    const sucsessForm = success.content.cloneNode(true);
-    document.body.appendChild(sucsessForm);
-    const sucsessBtn = document.querySelector('.success__button');
-    sucsessBtn.addEventListener('click', () => {
-      const success = document.querySelector('.success');
-      success.remove();
+
+     const formData = new FormData(evt.target);
+  fetch('https://27.javascript.pages.academy/kekstagram',
+    {
+      method: 'POST',
+      body: formData,
     });
-  }
+  const sucsessForm = success.content.cloneNode(true);
+  document.body.appendChild(sucsessForm);
+  const sucsessBtn = document.querySelector('.success__button');
+
+  sucsessBtn.addEventListener('click', () => {
+    hideSucsessWindow();
+  });
+
+  window.addEventListener('keydown', onPopupEscKeydown);
+
+  // document.addEventListener('click', () => {
+  //   hideSucsessWindow();
+  //   document.removeEventListener('click', hideSucsessWindow)
+  // });
+  } 
 });
